refactor: create cache_utils.py, move stuff there
This will eventually support the creation of a `content_metadata`
module, and we'll yank stuff out of the `subsidy_access_policy` module
into this new module.
