+++ enterprise_access/apps/subsidy_access_policy/utils.py
@@ -5,14 +5,8 @@ import hashlib
 
 from django.apps import apps
 from django.conf import settings
-from edx_django_utils.cache import RequestCache
 from edx_enterprise_subsidy_client import get_enterprise_subsidy_api_client
 from simple_history.models import HistoricalRecords, registered_models
-
-from enterprise_access import __version__ as code_version
-
-CACHE_KEY_SEP = ':'
-CACHE_NAMESPACE = 'subsidy_access_policy'
 
 LEDGERED_SUBSIDY_IDEMPOTENCY_KEY_PREFIX = 'ledger-for-subsidy'
 TRANSACTION_METADATA_KEYS = {
@@ -32,27 +26,6 @@ def get_versioned_subsidy_client():
     if getattr(settings, 'ENTERPRISE_SUBSIDY_API_CLIENT_VERSION', None):
         kwargs['version'] = int(settings.ENTERPRISE_SUBSIDY_API_CLIENT_VERSION)
     return get_enterprise_subsidy_api_client(**kwargs)
-
-
-def versioned_cache_key(*args):
-    """
-    Utility to produce a versioned cache key, which includes
-    an optional settings variable and the current code version,
-    so that we can perform key-based cache invalidation.
-    """
-    components = [str(arg) for arg in args]
-    components.append(code_version)
-    if stamp_from_settings := getattr(settings, 'CACHE_KEY_VERSION_STAMP', None):
-        components.append(stamp_from_settings)
-    decoded_cache_key = CACHE_KEY_SEP.join(components)
-    return hashlib.sha512(decoded_cache_key.encode()).hexdigest()
-
-
-def request_cache():
-    """
-    Helper that returns a namespaced RequestCache instance.
-    """
-    return RequestCache(namespace=CACHE_NAMESPACE)
 
 
 def create_idempotency_key_for_transaction(subsidy_uuid, **metadata):
