+++ enterprise_access/cache_utils.py
@@ -1,0 +1,33 @@
+"""
+Utils for interacting with cache interfaces.
+"""
+import hashlib
+
+from django.conf import settings
+from edx_django_utils.cache import RequestCache
+
+from enterprise_access import __version__ as code_version
+
+CACHE_KEY_SEP = ':'
+DEFAULT_NAMESPACE = 'enterprise-access-default'
+
+
+def versioned_cache_key(*args):
+    """
+    Utility to produce a versioned cache key, which includes
+    an optional settings variable and the current code version,
+    so that we can perform key-based cache invalidation.
+    """
+    components = [str(arg) for arg in args]
+    components.append(code_version)
+    if stamp_from_settings := getattr(settings, 'CACHE_KEY_VERSION_STAMP', None):
+        components.append(stamp_from_settings)
+    decoded_cache_key = CACHE_KEY_SEP.join(components)
+    return hashlib.sha512(decoded_cache_key.encode()).hexdigest()
+
+
+def request_cache(namespace=DEFAULT_NAMESPACE):
+    """
+    Helper that returns a namespaced RequestCache instance.
+    """
+    return RequestCache(namespace=namespace)
